implement usdc variant on mobile web
