+++ packages/web/src/components/track/PremiumConditionsPill.module.css
@@ -1,0 +1,36 @@
+.container {
+  align-items: center;
+  background-color: var(--accent-blue);
+  border-radius: var(--unit-1);
+  color: var(--static-white);
+  display: flex;
+  gap: var(--unit-1);
+  height: var(--unit-6);
+  padding: var(--unit-2) var(--unit-3);
+}
+
+.container svg {
+  width: var(--unit-4);
+  height: var(--unit-4);
+}
+
+.container path {
+  fill: var(--static-white);
+}
+
+.gatedContent {
+  background-color: var(--accent-blue);
+}
+
+.premiumContent {
+  background-color: var(--special-light-green);
+}
+
+.spinner {
+  height: var(--unit-4);
+  width: var(--unit-4);
+}
+
+.spinner g path {
+  stroke: var(--static-white);
+}
